Test Merge

See merge request s09-webmobile1-sub2/S09P12B303!26

+++ Server/src/main/java/com/threeracha/gaewoonhae/db/repository/UserRepository.java
@@ -10,5 +10,5 @@ import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
-    Optional<User> findByUserId(int userId);
+    Optional<User> findByUserId(Long userId);
 }
